Fix installation path of language files
Without this fix, this happens:

	$ make PREFIX=$HOME/nanotts
	$ make PREFIX=$HOME/nanotts install
	$ $HOME/nanotts/bin/nanotts -l en_GB -p 'hello there'
	Using Lingware directory: en_GB
	read: 12 bytes from command line
	using lang: en-GB
	Cannot load text analysis resource file (-40): cannot open file: en_GB/en-GB_ta.bin
	 * problem initializing Svox Pico

It's looking in a per-language subdirectory, but the files aren't installed in per
language subdirectories, so it doesn't find them.

This change makes it look in the correct directory in the install directory first,
and if it doesn't find them there, fall back to ./lang (via lingware_paths[0]) so that:

	$ make PREFIX=/usr/local
	$ ./nanotts -l en_GB -p 'hello'

still works.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>